ai-backend/requirements.txt
azure-functions
azure-core==1.35.0
azure-data-tables==12.7.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
isodate==0.7.2
multidict==6.6.4
propcache==0.3.2
requests==2.32.4
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
msal
PyJWT
